(PC-17806) refactor(auth): move useBeneficiaryValidationNavigation and useNextSubscriptionStep in helpers folder
