+++ src/features/auth/components/QuitSignupModal.native.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { QuitSignupModal } from 'features/auth/components/QuitSignupModal'
-import { SignupStep } from 'features/auth/signup/enums'
+import { SignupStep } from 'features/auth/enums'
 import { navigateToHome } from 'features/navigation/helpers'
 import { analytics } from 'libs/firebase/analytics'
 import { fireEvent, render } from 'tests/utils'
