+++ src/features/auth/enums.ts
