+++ src/features/auth/helpers/__mocks__/useBeneficiaryValidationNavigation.ts
