+++ src/features/auth/helpers/useNextSubscriptionStep.test.ts
@@ -2,12 +2,13 @@ import { rest } from 'msw'
 
 import { IdentityCheckMethod, NextSubscriptionStepResponse, SubscriptionStep } from 'api/gen'
 import { useAuthContext } from 'features/auth/context/AuthContext'
-import { useNextSubscriptionStep } from 'features/auth/signup/useNextSubscriptionStep'
 import { env } from 'libs/environment'
 import { useNetInfoContext as useNetInfoContextDefault } from 'libs/network/NetInfoWrapper'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { server } from 'tests/server'
 import { renderHook, waitFor } from 'tests/utils'
+
+import { useNextSubscriptionStep } from './useNextSubscriptionStep'
 
 const setError = jest.fn()
 jest.mock('features/auth/context/AuthContext')
