+++ src/features/auth/signup/AcceptCgu/AcceptCgu.tsx
@@ -4,7 +4,7 @@ import { v4 as uuidv4 } from 'uuid'
 import { CardContent, Paragraphe } from 'features/auth/components/signupComponents'
 import { useSettingsContext } from 'features/auth/context/SettingsContext'
 import { contactSupport } from 'features/auth/helpers/contactSupport'
-import { PreValidationSignupStepProps } from 'features/auth/signup/types'
+import { PreValidationSignupStepProps } from 'features/auth/types'
 import { env } from 'libs/environment'
 import { captureMonitoringError } from 'libs/monitoring'
 import { useNetInfoContext } from 'libs/network/NetInfoWrapper'
