+++ src/features/auth/signup/VerifyEligiblity/VerifyEligibility.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent, useState } from 'react'
 import { View } from 'react-native'
 import styled from 'styled-components/native'
 
-import { useBeneficiaryValidationNavigation } from 'features/auth/signup/useBeneficiaryValidationNavigation'
+import { useBeneficiaryValidationNavigation } from 'features/auth/helpers/useBeneficiaryValidationNavigation'
 import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ButtonTertiaryBlack } from 'ui/components/buttons/ButtonTertiaryBlack'
