+++ src/features/auth/types.ts
@@ -6,3 +6,17 @@ export type SignInResponseFailure = {
     general: string[]
   }
 }
+
+export type SignupData = {
+  email: string
+  marketingEmailSubscription: boolean
+  password: string
+  birthdate: string
+  postalCode: string
+}
+
+export interface PreValidationSignupStepProps {
+  accessibilityLabelForNextStep?: string
+  goToNextStep: (signupData: Partial<SignupData>) => void
+  signUp: (token: string) => Promise<void>
+}
