+++ src/features/eighteenBirthday/components/EighteenBirthdayCard.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 
+import { useBeneficiaryValidationNavigation } from 'features/auth/helpers/useBeneficiaryValidationNavigation'
 import { useDepositAmountsByAge } from 'features/auth/helpers/useDepositAmountsByAge'
-import { useBeneficiaryValidationNavigation } from 'features/auth/signup/useBeneficiaryValidationNavigation'
 import TutorialPassLogo from 'ui/animations/eighteen_birthday.json'
 import { AchievementCardKeyProps, GenericAchievementCard } from 'ui/components/achievements'
