+++ src/features/eighteenBirthday/pages/EighteenBirthday.web.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { useDepositAmountsByAge } from 'features/auth/helpers/useDepositAmountsByAge'
-import { useNextSubscriptionStep } from 'features/auth/signup/useNextSubscriptionStep'
+import { useNextSubscriptionStep } from 'features/auth/helpers/useNextSubscriptionStep'
 import { render, checkAccessibilityFor } from 'tests/utils/web'
 
 import { EighteenBirthday } from './EighteenBirthday'
@@ -10,7 +10,7 @@ jest.mock('features/auth/helpers/useDepositAmountsByAge')
 const mockUseDepositAmountsByAge = useDepositAmountsByAge as jest.Mock
 mockUseDepositAmountsByAge.mockReturnValue({ eighteenYearsOldDeposit: '300 €' })
 
-jest.mock('features/auth/signup/useNextSubscriptionStep')
+jest.mock('features/auth/helpers/useNextSubscriptionStep')
 const mockUseNextSubscriptionStep = useNextSubscriptionStep as jest.Mock
 mockUseNextSubscriptionStep.mockReturnValue({ navigateToNextBeneficiaryValidationStep: jest.fn() })
