+++ src/features/home/components/headers/HomeHeader.tsx
@@ -4,8 +4,8 @@ import { Platform } from 'react-native'
 import styled from 'styled-components/native'
 
 import { useAuthContext } from 'features/auth/context/AuthContext'
-import { useBeneficiaryValidationNavigation } from 'features/auth/signup/useBeneficiaryValidationNavigation'
-import { useNextSubscriptionStep } from 'features/auth/signup/useNextSubscriptionStep'
+import { useBeneficiaryValidationNavigation } from 'features/auth/helpers/useBeneficiaryValidationNavigation'
+import { useNextSubscriptionStep } from 'features/auth/helpers/useNextSubscriptionStep'
 import { GeolocationBanner } from 'features/home/components/GeolocationBanner'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { isUserBeneficiary } from 'features/profile/helpers/isUserBeneficiary'
