+++ src/features/identityCheck/api/api.ts
@@ -5,7 +5,7 @@ import { useMutation, useQuery } from 'react-query'
 import { api } from 'api/api'
 import { ApiError } from 'api/apiHelpers'
 import { MaintenancePageType, ProfileUpdateRequest, ProfileOptionsResponse } from 'api/gen'
-import { useNextSubscriptionStep } from 'features/auth/signup/useNextSubscriptionStep'
+import { useNextSubscriptionStep } from 'features/auth/helpers/useNextSubscriptionStep'
 import { useSubscriptionContext } from 'features/identityCheck/context/SubscriptionContextProvider'
 import { SubscriptionState } from 'features/identityCheck/context/types'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
