+++ src/features/identityCheck/pages/Stepper.web.test.tsx
@@ -12,7 +12,7 @@ import { IdentityCheckStepper } from './Stepper'
 
 mockdate.set(new Date('2020-12-01T00:00:00.000Z'))
 
-jest.mock('features/auth/signup/useNextSubscriptionStep')
+jest.mock('features/auth/helpers/useNextSubscriptionStep')
 
 jest.mock('features/identityCheck/useSetCurrentSubscriptionStep', () => ({
   useSetSubscriptionStepAndMethod: jest.fn(() => ({
