+++ src/features/identityCheck/pages/identification/ubble/IdentityCheckEnd.native.test.tsx
@@ -21,7 +21,7 @@ let mockNextSubscriptionStep: NextSubscriptionStepResponse = {
   nextSubscriptionStep: SubscriptionStep['honor-statement'],
 }
 
-jest.mock('features/auth/signup/useNextSubscriptionStep', () => ({
+jest.mock('features/auth/helpers/useNextSubscriptionStep', () => ({
   useNextSubscriptionStep: jest.fn(() => ({
     data: mockNextSubscriptionStep,
   })),
